codeexplorer: ask model to expore a codebase using tools
This aims to try out agent patterns. I saw Claude Code uses a very
striking methodology of giving the LLM the tools to explore a codebase
and then make changes --- you can see it doing various "human like"
things during the demo like listing directory content. This is in
contrast to other tools I've seen that tend to try to be smart
themselves, generating a sophisticated, context-laden prompt and
submitting to the model.

Let's try a simple agent, giving it tools to navigate the files in
a codebase and read them, with the goal of outputting an explanation
of what the codebase is for.