codeexplorer: provide tree directory listing tool
Replace the stack-based working directory approach with a single tool
that outputs the whole directory tree, like the `tree` command.

I'd observed that Haiku in particular had real trouble keeping track
of and navigating directories using the approach of a "working
directory" that it could change.

So I altered the code to allow for reading the whole directory tree
in one go such that the model didn't have to navigate around the
file system (although Sonnet 3.7 did okay with that). This seemed
to help Haiku a lot.

In addition, I'd seen that Haiku wasn't very good at understanding
that some tools took a `path` argument and some didn't --- it often
sent an argument for tools that didn't need one but looked like they
_could_ use on (eg, list_working_directory was often given a path
argument). This of course meant that Haiku assumed it was listing
one directory (via the path) but was getting confused by the returned
listing actually being in the current working directory. Watching the
read_file tool calls indicated this confusion, the paths were all
mangled.

Being more uniform with arguments for the tools that are similar,
taking paths for all of them, seems to help Haiku. That's an
interesting thing, that even relatively modern models like Haiku have
difficulty with tools.